// src/interfaces/dadata-api.ts
export namespace DadataApi {
// ...
  export type BaseRequest = {
    /**
     * Запрос, для которого нужно получить подсказки
     */
    query: string;
    /**
     * Количество возвращаемых подсказок (по умолчанию — 10, максимум — 20).
     */
    count?: number;
  };
// ...
  export interface Response<T> {
    /**
     * Отформатированные данные
     */
    value: string;
    /**
     * Неотформатированные данные
     */
    unrestricted_value: string;
    data: T;
  }
// ...
}